rename helper functions in timetrigger service for clarity
